requirements.txt
d3rlpy==1.1.0
gym==0.23.1
numpy==1.21.6
PyYAML==6.0
scikit_learn==1.0.2
torch==1.11.0
tqdm==4.64.0
mujoco_py==2.0.2.13
dm-control==0.0.295778102